Filemanager debug

+++ RosettaExportDeployment.inc.php
@@ -144,19 +144,25 @@ class RosettaExportDeployment
 				if (file_exists($xmlExport)) {
 					array_push($galleyFiles, $tmpExportFile);
 				}
-
+				$failedFiles = [];
 				foreach ($galleyFiles as $file) {
 
-					copy($file["fullFilePath"], join(DIRECTORY_SEPARATOR, array($STREAM_PATH, $file["path"], basename($file["fullFilePath"]))));
+					$copySuccess = copy($this->getPlugin()->getBasePath() . DIRECTORY_SEPARATOR . $file["fullFilePaths"], join(DIRECTORY_SEPARATOR, array($STREAM_PATH, $file["path"], basename($file["fullFilePath"]))));
+					if (!$copySuccess) $failedFiles [] = $this->getPlugin()->getBasePath() . DIRECTORY_SEPARATOR . $file["fullFilePath"];
 					foreach ($file["dependentFiles"] as $dependentFile) {
-						copy($dependentFile["fullFilePath"], join(DIRECTORY_SEPARATOR, array($STREAM_PATH, $file["path"], basename($dependentFile["fullFilePath"]))));
+						$copySuccess = copy($this->getPlugin()->getBasePath() . DIRECTORY_SEPARATOR . $dependentFile["fullFilePath"], join(DIRECTORY_SEPARATOR, array($STREAM_PATH, $file["path"], basename($dependentFile["fullFilePath"]))));
+						if (!$copySuccess) $failedFiles [] = $this->getPlugin()->getBasePath() . DIRECTORY_SEPARATOR . $dependentFile["fullFilePath"];
 					}
 				}
 
 				exec('java -jar ' . $this->getPlugin()->getPluginPath() . '/bin/xsd11-validator.jar -if ' . $IE_PATH . ' -sf ' . $this->getPlugin()->getPluginPath() . '/schema/mets_rosetta.xsd ', $validationOutPut, $validationStatus);
-				if (!$isTest and $validationStatus == 0) {
+				if (!$isTest and $validationStatus == 0 && count($failedFiles) == 0) {
 					$this->doDeposit($context, $INGEST_PATH, $SIP_PATH, $submission);
 					unlink($xmlExport);
+				} else {
+					foreach ($failedFiles as $failedFile) {
+						var_dump("Copy failed: " . $failedFile);
+					}
 				}
